+++ tests/podmander-controller-service-json_utils_tests.ads
@@ -1,0 +1,11 @@
+--  Copyright (C) 2026 Jochen Lillich
+--  SPDX-License-Identifier: Apache-2.0
+
+with AUnit;
+with AUnit.Test_Suites;
+
+package Podmander.Controller.Service.Json_Utils_Tests is
+
+   function Suite return AUnit.Test_Suites.Access_Test_Suite;
+
+end Podmander.Controller.Service.Json_Utils_Tests;